Remove Compiler bugs
