+++ ws/src/main/scala/de/thm/ii/fbs/config/WebSocketConfig.scala
@@ -2,8 +2,8 @@ package de.thm.ii.fbs.config
 
 import java.security.Principal
 
-import de.thm.ii.fbs.model.practiceroom.UserSessionMap
-import de.thm.ii.fbs.services.persistance.UserService
+import de.thm.ii.fbs.model.classroom.UserSessionMap
+import de.thm.ii.fbs.services.security.AuthService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.{Bean, Configuration, Primary}
 import org.springframework.messaging.simp.SimpMessageType
@@ -24,11 +24,13 @@ import org.springframework.web.socket.messaging._
 class WebSocketConfig extends WebSocketMessageBrokerConfigurer {
   private val registry: DefaultSimpUserRegistry = new DefaultSimpUserRegistry();
   private val resolver: DefaultUserDestinationResolver = new DefaultUserDestinationResolver(registry);
+
   @Autowired
-  private val userService: UserService = null
+  private val authService: AuthService = null
   /**
     * @return Default user registry.
     */
+
   @Bean
   @Primary
   def userRegistry(): SimpUserRegistry = registry
@@ -36,6 +38,7 @@ class WebSocketConfig extends WebSocketMessageBrokerConfigurer {
   /**
     * @return Default user destination resolver.
     */
+
   @Bean
   @Primary
   def userDestinationResolver(): UserDestinationResolver = resolver
@@ -62,23 +65,23 @@ class WebSocketConfig extends WebSocketMessageBrokerConfigurer {
     * @param registration Hockup on the client inbound channel.
     */
   override def configureClientInboundChannel(registration: ChannelRegistration): Unit = {
-    registration.interceptors(new JWTAuthenticationChannelInterceptor(registry, userService))
+    registration.interceptors(new JWTAuthenticationChannelInterceptor(registry, authService))
   }
 }
 
 /**
   * JWT authentication
   * @param registry User registry
-  * @param userService User service
+  * @param authService Authentication service
   */
-class JWTAuthenticationChannelInterceptor(registry: DefaultSimpUserRegistry, userService: UserService) extends ChannelInterceptor {
+class JWTAuthenticationChannelInterceptor(registry: DefaultSimpUserRegistry, authService: AuthService) extends ChannelInterceptor {
   /**
     * Hockup for jwt authentication.
     * @param message Message
     * @param channel Channel
     * @return Passed messages
     */
-  override def preSend(message: Message[_], channel: MessageChannel): Message[_] = {
+   override def preSend(message: Message[_], channel: MessageChannel): Message[_] = {
     val accessor = StompHeaderAccessor.wrap(message)
     if (accessor.getMessageType == SimpMessageType.DISCONNECT) {
       UserSessionMap.delete(accessor.getSessionId)
@@ -91,18 +94,13 @@ class JWTAuthenticationChannelInterceptor(registry: DefaultSimpUserRegistry, use
         throw new MessagingException("Not authenticated!")
       }
 
-      val userOpt = userService.verifyUserByTocken(jwtToken)
-      if (userOpt.isEmpty) {
-        throw new MessagingException("Not authenticated: Invalid token.")
-      }
+      val user = authService.authorize(jwtToken)
 
-      val principal: Principal = userOpt.get
-
-      accessor.setUser(principal)
+      accessor.setUser(user)
       accessor.setLeaveMutable(true)
 
       if (accessor.getMessageType == SimpMessageType.CONNECT) {
-        UserSessionMap.map(accessor.getSessionId, principal)
+        UserSessionMap.map(accessor.getSessionId, user)
       }
 
       if (accessor.getMessageType match {
@@ -110,9 +108,9 @@ class JWTAuthenticationChannelInterceptor(registry: DefaultSimpUserRegistry, use
         case _ => false
       }) {
         registry.onApplicationEvent(accessor.getMessageType match {
-          case SimpMessageType.CONNECT => new SessionConnectedEvent(this, message.asInstanceOf[Message[Array[Byte]]], principal)
-          case SimpMessageType.SUBSCRIBE => new SessionSubscribeEvent(this, message.asInstanceOf[Message[Array[Byte]]], principal)
-          case SimpMessageType.UNSUBSCRIBE => new SessionUnsubscribeEvent(this, message.asInstanceOf[Message[Array[Byte]]], principal)
+          case SimpMessageType.CONNECT => new SessionConnectedEvent(this, message.asInstanceOf[Message[Array[Byte]]], user)
+          case SimpMessageType.SUBSCRIBE => new SessionSubscribeEvent(this, message.asInstanceOf[Message[Array[Byte]]], user)
+          case SimpMessageType.UNSUBSCRIBE => new SessionUnsubscribeEvent(this, message.asInstanceOf[Message[Array[Byte]]], user)
         })
       }
       MessageBuilder.createMessage(message.getPayload, accessor.getMessageHeaders)
