+++ ws/src/main/scala/de/thm/ii/fbs/controller/CheckerConfigurationController.scala
@@ -5,8 +5,6 @@ import java.nio.file.Files
 import com.fasterxml.jackson.databind.JsonNode
 import de.thm.ii.fbs.controller.exception.{BadRequestException, ForbiddenException, ResourceNotFoundException}
 import de.thm.ii.fbs.model.{CheckrunnerConfiguration, CourseRole, GlobalRole}
-import de.thm.ii.fbs.services.StorageService
-import de.thm.ii.fbs.services.core.StorageService
 import de.thm.ii.fbs.services.persistance.{CheckerConfigurationService, CourseRegistrationService, StorageService}
 import de.thm.ii.fbs.services.security.AuthService
 import de.thm.ii.fbs.util.JsonWrapper.jsonNodeToWrapper
