+++ ws/src/main/scala/de/thm/ii/fbs/model/classroom/storage/Index.scala
@@ -1,4 +1,4 @@
-package de.thm.ii.fbs.model.practiceroom.storage
+package de.thm.ii.fbs.model.classroom.storage
 
 import scala.collection.mutable
